[FLINK-20303] Refactor out the test driver
